BGDIINF_SB-1484: Own QueryClass for fiter functions
Putting general filter functions to become methods of a custom QuerySet class
