+++ app/stac_api/managers.py
@@ -1,0 +1,163 @@
+import logging
+from datetime import datetime
+
+from django.contrib.gis.db import models
+from django.contrib.gis.geos import Point
+from django.contrib.gis.geos import Polygon
+from django.db.models import Q
+from django.utils.translation import gettext_lazy as _
+
+from rest_framework.exceptions import ValidationError
+
+from stac_api.validators import validate_geometry
+
+logger = logging.getLogger(__name__)
+
+
+class ItemQuerySet(models.QuerySet):
+
+    def by_bbox(self, bbox):
+        '''Filter a querystring with a given bbox
+
+        This function is a helper function, for some views, to add a bbox filter to the queryset.
+
+        Args:
+            queryset:
+                A django queryset (https://docs.djangoproject.com/en/3.0/ref/models/querysets/)
+            bbox:
+                A string defining a spatial bbox (f.ex. 5.96, 45.82, 10.49, 47.81)
+
+        Returns:
+            The queryset with the added spatial filter
+
+        Raises:
+            ValidationError: When the bbox does not contain 4 values. Or when the polygon build
+            from the bbox string is invalid.
+        '''
+        try:
+            logger.debug('Query parameter bbox = %s', bbox)
+            list_bbox_values = bbox.split(',')
+            if (
+                list_bbox_values[0] == list_bbox_values[2] and
+                list_bbox_values[1] == list_bbox_values[3]
+            ):
+                bbox_geometry = Point(float(list_bbox_values[0]), float(list_bbox_values[1]))
+            else:
+                bbox_geometry = Polygon.from_bbox(list_bbox_values)
+            validate_geometry(bbox_geometry)
+
+        except (ValueError, ValidationError, IndexError) as error:
+            logger.error(
+                'Invalid bbox query parameter: '
+                'Could not transform bbox "%s" to a polygon; %s'
+                'f.ex. bbox=5.96, 45.82, 10.49, 47.81',
+                bbox,
+                error
+            )
+            raise ValidationError(
+                _('Invalid bbox query parameter, '
+                  ' has to contain 4 values. f.ex. bbox=5.96,45.82,10.49,47.81'),
+                code='bbox-invalid'
+            )
+
+        return self.filter(geometry__intersects=bbox_geometry)
+
+    def by_datetime(self, date_time):
+        '''Filter a queryset by datetime
+
+        Args:
+            queryset:
+                 A django queryset (https://docs.djangoproject.com/en/3.0/ref/models/querysets/)
+            date_time:
+                A string
+        Returns:
+            The queryset filtered by date_time
+        '''
+        start, end = self._parse_datetime_query(date_time)
+        if end is not None:
+            return self._filter_by_datetime_range(start, end)
+        return self.filter(properties_datetime=start)
+
+    def _filter_by_datetime_range(self, start_datetime, end_datetime):
+        '''Filter a queryset by datetime range
+
+        Helper function of filter_by_datetime
+
+        Args:
+            queryset:
+                A django queryset (https://docs.djangoproject.com/en/3.0/ref/models/querysets/)
+            start_datetime:
+                A string with the start datetime
+            end_datetime:
+                A string with the end datetime
+        Returns:
+            The queryset filtered by datetime range
+        '''
+        if start_datetime == '..':
+            # open start range
+            return self.filter(
+                Q(properties_datetime__lte=end_datetime) |
+                Q(properties_end_datetime__lte=end_datetime)
+            )
+        if end_datetime == '..':
+            # open end range
+            return self.filter(
+                Q(properties_datetime__gte=start_datetime) |
+                Q(properties_end_datetime__gte=start_datetime)
+            )
+            # else fixed range
+        return self.filter(
+            Q(properties_datetime__range=(start_datetime, end_datetime)) | (
+                Q(properties_start_datetime__gte=start_datetime) &
+                Q(properties_end_datetime__lte=end_datetime)
+            )
+        )
+
+    def _parse_datetime_query(self, date_time):
+        '''Parse the datetime query as specified in the api-spec.md.
+
+        Returns one of the following
+            datetime, None
+            datetime, '..'
+            '..', datetime
+            datetime, datetime
+        '''
+        start, sep, end = date_time.partition('/')
+        try:
+            if start != '..':
+                start = datetime.fromisoformat(start.replace('Z', '+00:00'))
+            if end and end != '..':
+                end = datetime.fromisoformat(end.replace('Z', '+00:00'))
+        except ValueError as error:
+            logger.error(
+                'Invalid datetime query parameter "%s", must be isoformat; %s', date_time, error
+            )
+            raise ValidationError(
+                _('Invalid datetime query parameter, must be isoformat'),
+                code='datetime'
+            )
+
+        if end == '':
+            end = None
+
+        if start == '..' and (end is None or end == '..'):
+            logger.error(
+                'Invalid datetime query parameter "%s"; '
+                'cannot start with open range when no end range is defined',
+                date_time
+            )
+            raise ValidationError(
+                _('Invalid datetime query parameter, '
+                  'cannot start with open range when no end range is defined'),
+                code='datetime'
+            )
+        return start, end
+
+
+class ItemManager(models.Manager):
+
+    def get_queryset(self):
+        return ItemQuerySet(self.model, using=self._db).select_related('collection')
+
+    def by_bbox(self, bbox):
+        return self.get_queryset().by_bbox(bbox)
